requirements.txt
biopython~=1.81
typing-extensions~=4.4.0
